requirements.txt
aioice==0.9.0
aiortc==1.10.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
asyncgui==0.6.3
asynckivy==0.6.4
audio-processing==0.0.1
av==13.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
Cython==3.0.12
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
exceptiongroup==1.2.2
faiss-cpu==1.10.0
filelock==3.17.0
filetype==1.2.0
fonttools==4.56.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.31.0
google-crc32c==1.6.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip#sha256=ab18d4cc36e2d53ea87f463d01299c39c75bae66cc1950f9bc6397f23317c848
kiwisolver==1.4.8
MarkupSafe==3.0.2
materialyoucolor==2.0.10
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
noisereduce==3.0.3
numpy==2.2.3
olefile==0.47
openai==1.63.2
packaging==24.2
pdf2image==1.17.0
pefile==2023.2.7
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pylibsrtp==0.11.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypiwin32==223
pyroomacoustics==0.8.3
pystray==0.19.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
six==1.17.0
passlib
sniffio==1.3.1
sounddevice==0.5.1
stripe==11.6.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
webrtcvad==2.0.10
